backend/requirements.txt
openai
pinecone
voyageai
python-dotenv
